Initialize recipe card UI
Recipe generator now pulls multiple details from the API request,
including ingredients, preparation time, and a link to the original
recipe, and displays the information on an MUI card. Additional
formatting still needs to be done, significantly so on the placeholder
card which renders prior to first GET request. Query options/UI is also
on the to-do list.